pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wagtail-amazon-paapi"
version = "0.1.0"
authors = [
  { name="Ravi Gupta", email="ravi.opensource@protonmail.com" },
]
description = "A Wagtail package for managing and displaying Amazon products using Amazon PA API 5."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Wagtail",
    "Framework :: Django",
]
dependencies = [
    "django>=5.2",
    "wagtail>=7.0.1",
    "python-amazon-paapi>=5.0.1",
]